+++ packages/message-bus/src/index.ts
@@ -1,2 +1,2 @@
 export * from './types';
-export { createMessageBus, type MessageBus } from './message-bus';
+export * from './message-bus';