+++ test-config.ts
@@ -1,0 +1,3 @@
+export default {
+  plugins: ['@auto-engineer/server-implementer'],
+};