+++ packages/frontend-generator-react-graphql/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @auto-engineer/frontend-react-graphql-generator
 
+## 0.1.3
+
+### Patch Changes
+
+- version test
+- Updated dependencies
+  - @auto-engineer/message-bus@0.5.4
+
+## 0.1.2
+
+### Patch Changes
+
+- fix version report
+- Updated dependencies
+  - @auto-engineer/message-bus@0.5.3
+
 ## 0.1.1
 
 ### Patch Changes